split demo examples by packages
