+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint --ignore-path=.eslintignore --ext .js,.jsx ./packages/",
+    "publish": "npm run lint && lerna publish && lerna run gh-pages",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": "https://github.com/williaster/data-ui.git",
@@ -33,6 +34,7 @@
     "react-dom": "~15.4.2"
   },
   "peerDependencies": {
+    "lerna": "2.0.0-rc.2",
     "react": "~15.4.2",
     "react-dom": "~15.4.2"
   }
