+++ packages/demo/package.json
@@ -25,6 +25,7 @@
     "aphrodite": "^1.2.0",
     "babel-preset-airbnb": "^2.2.3",
     "git-directory-deploy": "^1.5.1",
+    "glob": "^7.1.1",
     "immutable": "^3.8.1",
     "react": "~15.4.2",
     "react-dom": "~15.4.2",
